Files clenaed up and controllers extracted from server file
