+++ app/controllers/sessions.rb
@@ -1,0 +1,24 @@
+get '/sessions/new' do
+  if session[:user_id]
+    redirect to ('/')
+  end
+  erb :"sessions/new"
+end
+
+post '/sessions' do
+  email, password = params[:email], params[:password]
+  user = User.authenticate(email, password)
+  if user
+    session[:user_id] = user.id
+    redirect to('/')
+  else
+    flash[:errors] = ["The email or password is incorrect"]
+    erb :"sessions/new"
+  end
+end
+
+delete '/sessions' do
+  flash[:notice] = "Good bye!"
+  session[:user_id] = nil
+  erb :"sessions/new"
+end
