+++ spec/features/listing_all_peep_spec.rb
@@ -2,10 +2,10 @@ require 'spec_helper'
 
 feature "user browses the list of peeps" do
 
-  before(:each) {
+  before do
     Peep.create(:content => "This is the first chitter peep",
                 :created_at => DateTime.now)
-  }
+  end
 
   scenario " when opening the home page" do
     visit '/'
